refactor: update import paths for updateSongId3Tags and related functions
