+++ .vscode/settings.json
@@ -81,7 +81,7 @@
   },
   "editor.tokenColorCustomizations": {
     "comments": "#657bae",
-    "textMateRules": []
+    "textMateRules": [ ]
   },
   "editor.wordWrap": "off",
   "window.zoomLevel": -1,
