+++ src/main/core/resolveFeaturingArtists.ts
@@ -1,4 +1,4 @@
-import updateSongId3Tags from '../updateSongId3Tags';
+import updateSongId3Tags from '../updateSong/updateSongId3Tags';
 import sendSongID3Tags from './sendSongId3Tags';
 import { getArtistsData } from '../filesystem';
 import logger from '../logger';
