+++ src/main/core/resolveSeparateArtists.ts
@@ -1,6 +1,6 @@
 import { getArtistArtworkPath } from '../fs/resolveFilePaths';
 import logger from '../logger';
-import updateSongId3Tags from '../updateSongId3Tags';
+import updateSongId3Tags from '../updateSong/updateSongId3Tags';
 import {
   getAlbumsData,
   getArtistsData,
